+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Changes to Calva.
 
 ## [Unreleased]
 
+## [2.0.135 - 2020-12-20]
+- [Binding keys to REPL functions, passing the namespace and cursor line (Notespace integration)](https://github.com/BetterThanTomorrow/calva/issues/863)
+- [Make REPL prompt submit if the cursor is after the top level form](https://github.com/BetterThanTomorrow/calva/issues/875)
+- [Only print stacktrace on demand](https://github.com/BetterThanTomorrow/calva/issues/878)
+
 ## [2.0.134 - 2020-12-05]
 - Fix: [Live share jackout error](https://github.com/BetterThanTomorrow/calva/issues/856)
 - Fix: [Cannot read property 'document' of undefined](https://github.com/BetterThanTomorrow/calva/issues/846)